实现友好的emit和debug check

+++ src/main.cpp
@@ -5,8 +5,7 @@
 #include "Geometry.h"
 #include "Run.h"
 
-
-int main(int argc, char *argv[])
+int main(int argc, char* argv[])
 {
     Kokkos::ScopeGuard scope_guard(argc, argv);
     Kokkos::printf("DefaultExecutionSpace: %s HostSpace: %s\n", ExecSpace::name(), Kokkos::HostSpace::name());
@@ -14,8 +13,9 @@ int main(int argc, char *argv[])
 
     Run run("data/MultiLayers.vol");
     auto res = run.run(1);
-    for(int i = 0; i < 1; i++){
-        printf("%d\n", res(i).type);
+    for (int i = 0; i < 1; i++)
+    {
+        printf("%d\n", (int)res(i).type);
     }
     return 0;
 }
